refactor: Rename runner gprc `spec` -> `runner`
