+++ runner/protos/runner.proto
@@ -1,5 +1,5 @@
 syntax = "proto3";
-package spec;
+package runner;
 
 service Runner {
     // Starts a new Runner executor
