+++ runner/src/server/runner-service.ts
@@ -1,12 +1,12 @@
 import { type ServerUnaryCall, type sendUnaryData } from '@grpc/grpc-js';
-import { type RunnerHandlers } from '../generated/spec/Runner';
-import { type StartExecutorResponse__Output, type StartExecutorResponse } from '../generated/spec/StartExecutorResponse';
-import { type StartExecutorRequest__Output } from '../generated/spec/StartExecutorRequest';
-import { type StopExecutorRequest__Output } from '../generated/spec/StopExecutorRequest';
-import { type StopExecutorResponse__Output, type StopExecutorResponse } from '../generated/spec/StopExecutorResponse';
-import { type ListExecutorsRequest__Output } from '../generated/spec/ListExecutorsRequest';
-import { type ListExecutorsResponse__Output, type ListExecutorsResponse } from '../generated/spec/ListExecutorsResponse';
-import { type ExecutorInfo__Output } from '../generated/spec/ExecutorInfo';
+import { type RunnerHandlers } from '../generated/runner/Runner';
+import { type StartExecutorResponse__Output, type StartExecutorResponse } from '../generated/runner/StartExecutorResponse';
+import { type StartExecutorRequest__Output } from '../generated/runner/StartExecutorRequest';
+import { type StopExecutorRequest__Output } from '../generated/runner/StopExecutorRequest';
+import { type StopExecutorResponse__Output, type StopExecutorResponse } from '../generated/runner/StopExecutorResponse';
+import { type ListExecutorsRequest__Output } from '../generated/runner/ListExecutorsRequest';
+import { type ListExecutorsResponse__Output, type ListExecutorsResponse } from '../generated/runner/ListExecutorsResponse';
+import { type ExecutorInfo__Output } from '../generated/runner/ExecutorInfo';
 import type StreamHandler from '../stream-handler';
 import * as grpc from '@grpc/grpc-js';
 import assert from 'assert';
